se realizo el engineering task S08_T1

+++ FoodYeah/Dto/CardDto.cs
@@ -5,10 +5,10 @@ namespace FoodYeah.Dto
     public class CardCreateDto
     {
         [Required]
-        public int CardNumber { get; set; }
+        public long CardNumber { get; set; }
         public int CustomerId { get; set; }
         public bool CardType { get; set; }
-        public byte CardCvi  { get; set; }
+        public int CardCvi  { get; set; }
         public string CardExpireDate { get; set; }
     }
 
@@ -20,11 +20,11 @@ namespace FoodYeah.Dto
     public class CardDto
     {
         public int CardId { get; set; }
-        public int CardNumber { get; set; }
+        public long CardNumber { get; set; }
         public int CustomerId { get; set; }
         public CustomerDto Customer { get; set; }
         public bool CardType { get; set; }
-        public byte CardCvi  { get; set; }
+        public int CardCvi  { get; set; }
         [Required]
         public string CardOwnerName { get; set; }
         public string CardExpireDate { get; set; }
@@ -33,10 +33,10 @@ namespace FoodYeah.Dto
      public class CardSimpleDto
     {
         public int CardId { get; set; }
-        public int CardNumber { get; set; }
+        public long CardNumber { get; set; }
         public int CustomerId { get; set; }
         public bool CardType { get; set; }
-        public byte CardCvi  { get; set; }
+        public int CardCvi  { get; set; }
         [Required]
         public string CardOwnerName { get; set; }
         public string CardExpireDate { get; set; }
